integrated ReadWrite component logic into AddItem and List files
