+++ src/components/AddItem.js
@@ -1,7 +1,32 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { withFirestore } from 'react-firestore';
 
-const AddItem = () => {
-  return <div>ADD AN ITEM</div>;
+const AddItem = props => {
+  const { firestore } = props;
+  const [enteredValue, setEnteredValue] = useState();
+
+  const addMsg = e => {
+    e.preventDefault();
+    firestore.collection('example').add({
+      msg: enteredValue,
+    });
+  };
+  return (
+    <>
+      <form>
+        <input
+          title="message"
+          type="text"
+          name="msg"
+          placeholder="enter your message here..."
+          onChange={e => setEnteredValue(e.target.value)}
+        />
+        <button type="submit" onClick={e => addMsg(e)}>
+          Submit
+        </button>
+      </form>
+    </>
+  );
 };
 
-export default AddItem;
+export default withFirestore(AddItem);
