+++ src/components/List.js
@@ -1,7 +1,28 @@
 import React from 'react';
+import { FirestoreCollection, withFirestore } from 'react-firestore';
 
 const List = () => {
-  return <div>LIST</div>;
+  return (
+    <>
+      <FirestoreCollection
+        path="example"
+        render={({ isLoading, data }) => {
+          return isLoading ? (
+            <p>loading...</p>
+          ) : (
+            <div>
+              <h1>Messages</h1>
+              <ul style={{ listStyleType: 'none' }}>
+                {data.map(msg => (
+                  <li key={msg.id}>{msg.msg}</li>
+                ))}
+              </ul>
+            </div>
+          );
+        }}
+      />
+    </>
+  );
 };
 
 export default List;
